requirements.txt
Bootstrap_Flask==2.3.3
Flask==2.3.2
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.25
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv~=1.0.1
